update desi pong game
add pause, resume and reset game functionality. "l" for pause, "j" for resume and "d" for reset game.
now the game has high score limit and this is 10.
after reaching the score 10 by player 1 or player 2 the game will stop and display the player 1/ player 2 is win and game over text on the screen.
Thank You.