fix: Fix the syntax highlighting problem.
I used git bisect to determine that commit 2665941 was causing the
problem with syntax highlighting. It put the yarn.lock into the
repository, which is good practice for applications. I deleted the
yarn.lock file and rebuilt it on the main branch. Syntax highlighting
worked. It did update four packages.

There was also a warning about a dependency for lezer/common not being
met, which I fixed by adding it to package.json.

After fixing the warning, I tried running the original lock file, but
the bug was still present. So hopefully the rebuilt lock file fixes
things. We'll see when it's deployed.