Fire Hound by Himir Desai
A Fire hound is desperate to get to his home but countless portals are to be crossed in the way. Rain is a big obstacle for him and will instantly kill him if it touches the drop
Play this game with your friends chance by chance and score the Highest among your friends. The score is simply the number of portals passed by the hound during your turn.