Edited the Code
Fixed the glitch of extra carrots spawning by clearing the whole screen after a level reset. I also fixed the errors given by clicking the buttons after the game is over by moving the Bunny to 0,0 and not moving the Bunny or advancing any steps if the steps left is less than or equal to 0.