Add files via upload
This the which I have made with the help of my friends and we really worked hard on it. This game is about Flipping a Coin like you do in real life but the only difference is that you can flipping either on your Sprig or on your computer. With this you can play game like truth and dare like based on what you get you can play truth or dare with you friends.