Title
This game is an artistic, one which only few can truly grasp. 24hrs of LeStoopid places you directly into the drivers seat of a Le Mans driver, for a truly staggering 24hrs straight. Now some may question the authenticity of this experience, and let me put the to rest. I am both a veteran Le Mans driver, and decorated artist, and let me tell you, this game is the REAL DEAL!