feat: send a new copy of the code to the server only when the code changes
previously the client chat feature will send 2 requests to the llm api. the first being a copy of the current code and then the chat message sent by the user.

with these changes, we hash the code and only send a new copy of the code to the server when the hash of the current code in the editor is different from the previous one.