wall_defense (#3308)
* Add files via upload

* Rename Wave game .js to wave_game.js

* wall_defense

the finished version of wave game

* wall_defense

an image of the gameplay

* Add files via upload

* Delete games/wave_game.js

the updated version of wave_game has arrived. Now it is called wall-defense

* Delete games/img/Screenshot 2025-06-04 091053.png

I'm removing the screenshot. So that I can reupload it with the correct name

* wall_defense

screenshots of the gameplay and the menu

* Delete games/wall_defense .js

name of file does not fullfill metadata requests

* name change of game

* Delete games/img/Screenshot 2025-06-04 091053.png

name does not fulfill the metadata request

* Delete games/img/Screenshot 2025-09-20 153126.png

name does not fullfill the metadata request

* renamed menu image

* Delete games/wall_defense.js

I might have modified the coment at the start of the code.

* modified the comment at the start of the code

* Delete games/wall_defense .js

I forgot to remove the space from the original code. My bad :(

* removed space

The error was inthe original code in the editor. My bad :(

---------

Co-authored-by: graham <graham@hackclub.com>